[Feat] Inoperationid 바탕으로 정보 불러오기

+++ AOS/app/src/main/java/org/softeer/robocar/data/repository/onboard/OnboardRepository.kt
@@ -6,6 +6,6 @@ import org.softeer.robocar.data.model.Onboard
 
 interface OnboardRepository {
     suspend fun getOnboardDetails(
-        inOperationId: Int
+        inOperationId: Long
     ): Result<OnboardData>
 }
